--- src/components/trade/PriceEstimations/index.tsx
@@ -1,2 +1,0 @@
-export { PriceEstimations } from './PriceEstimations'
-export { PriceEstimationsComp } from './PriceEstimationComp'
