+++ src/components/trade/PriceSuggestions/index.tsx
@@ -1,0 +1,2 @@
+export { PriceSuggestions } from './PriceSuggestions'
+export { PriceSuggestionsComp } from './PriceSuggestionComp'
